legacy/requirements.txt
crewai==0.100.0
langchain>=0.0.325
langchain-community>=0.0.1
langchain-ollama>=0.0.1
PyYAML==6.0.1
python-dotenv==1.0.1
rich>=13.7.0